+++ R/data.R
@@ -1,6 +1,6 @@
 #' Example soil XRPD data
 #'
-#' 3 soil samples from different parent materials measured by XRPD
+#' 3 soil samples from different parent materials measured by XRPD (Cu K-alpha radiation)
 #'
 #' @format A list of 3 dataframes (named according to rock type),
 #' with each dataframe containing two columns of:
@@ -13,8 +13,9 @@
 #' An example powdRlib reference library
 #'
 #' This \code{powdRlib} object, built using the \code{powdRlib} constructor function,
-#' contains a range of measured XRPD data along with their reference intensity ratios.
-#' The library can be used with the \code{soils} example data for full pattern summation.
+#' contains a range of measured XRPD data (Cu K-alpha radiation) along with their reference
+#' intensity ratios. The library can be used with the \code{soils} example data for full
+#' pattern summation.
 #'
 #' @format A list of 3
 #' \describe{
@@ -61,10 +62,11 @@
 #' RockJock reference library
 #'
 #' A \code{powdRlib} object of 168 pure reference patterns from the RockJock
-#' library along with reference intensity ratios. Can be used with the \code{fps()}
-#' and \code{afps()} functions for quantitative analysis. Example mixtures for
-#' testing the \code{rockjock} library with known concentrations are available
-#' in the \code{rockjock_mixtures} data. See \code{?rockjock_mixtures}.
+#' library (Cu K-alpha radiation) along with reference intensity ratios.Can be
+#' used with the \code{fps()} and \code{afps()} functions for quantitative
+#' analysis. Example mixtures for testing the \code{rockjock} library with known
+#' concentrations are available in the \code{rockjock_mixtures} data. See
+#' \code{?rockjock_mixtures}.
 #'
 #' @format A list of 3 components
 #' \describe{
@@ -81,8 +83,12 @@
 
 #' RockJock synthetic mixtures
 #'
-#' A list containing 8 XRPD measurements of synthetic mixtures that can be used to
-#' assess accuracy of quantitative analysis from the \code{fps()} and \code{afps()} functions.
+#' A list containing 8 XRPD measurements (Cu K-alpha radiation) of synthetic mixtures
+#' that can be used to assess accuracy of quantitative analysis from the \code{fps()}
+#' and \code{afps()} functions. The mixtures contain various amounts of quartz (QUARTZ
+#' standard in of the \code{rockjock} library), K-feldspar (ORDERED_MICROCLINE),
+#' plagioclase (LABRADORITE), kaolinite (KAOLINITE_DRY_BRANCH), dioctahedral smectite
+#' (MONTMORILLIONITE_WYO), illite (ILLITE_1M_RM30) and corundum (CORUNDUM).
 #'
 #' @format A list of 8 components, each comprised of two columns. Column \code{tth} specifies
 #' the 2theta axis and \code{counts} specifies the count intensities